+++ src/Kconfig
@@ -80,8 +80,11 @@ endmenu
 # Generic configuration options for CANbus
 config CANSERIAL
     bool
+config CANBUS
+    bool
+    default y if CANSERIAL
 config CANBUS_FREQUENCY
-    int "CAN bus speed" if LOW_LEVEL_OPTIONS && CANSERIAL
+    int "CAN bus speed" if LOW_LEVEL_OPTIONS && CANBUS
     default 500000
 config CANBUS_FILTER
     bool
