+++ src/generic/canbus.h
@@ -3,9 +3,6 @@
 
 #include <stdint.h> // uint32_t
 
-#define CANBUS_ID_ADMIN 0x3f0
-#define CANBUS_ID_ADMIN_RESP 0x3f1
-
 struct canbus_msg {
     uint32_t id;
     uint32_t dlc;
@@ -27,6 +24,5 @@ void canbus_set_filter(uint32_t id);
 // canbus.c
 void canbus_notify_tx(void);
 void canbus_process_data(struct canbus_msg *msg);
-void canbus_set_uuid(uint8_t *raw_uuid, uint32_t raw_uuid_len);
 
 #endif // canbus.h
